requirements.txt
bleach==2.1.3
boto3==1.7.8
botocore==1.10.8
Django==2.0.4
django-bleach==0.3.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-storages==1.6.6
django-wysiwyg==0.8.0
djangorestframework==3.8.2
docutils==0.14
gunicorn==19.8.1
html5lib==1.0.1
jmespath==0.9.3
Pillow==5.1.0
psycopg2==2.7.4
python-dateutil==2.7.2
pytz==2018.3
s3transfer==0.1.13
six==1.11.0
webencodings==0.5.1
